feat: add bot initialization tasks for Celery
